Fix powershell payload crashing upon command input.
When the reverse connection is received from the victim powershell, SET will prompt for shell command. Anything that is entered will trigger a crash:
"Something went wrong, printing the error: a bytes-like object is required, not 'str'"

By encoding the input string to bytes, it allows the program to continue.